requirements.txt
discord.py
aiohttp
customtkinter
